Extract editor transclusion features
