+++ apps/editor/src/App.tsx
@@ -39,6 +39,7 @@ import type { AppPhase, ConnectionState, ContractCatalogLoadState, CreationConte
 import { gatewayError, missingCoreCapabilities, missingTypeCapabilities } from "./gateway";
 import { useCollectionAuthorization } from "./collection-authorization";
 import { OpeningScreen, TypeWorkspaceLoading } from "./LoadingScreens";
+import { MarkdownNoteEditor } from "./MarkdownNoteEditor";
 import { backlinksFor, linkSuggestions, unresolvedNoteTarget } from "./links";
 import {
   COLLECTION_WIDTH,
@@ -94,6 +95,7 @@ import { useFileInventory } from "./use-file-inventory";
 import { useFileAssetStore } from "./use-file-assets";
 import { useFileWorkspace } from "./use-file-workspace";
 import { useEmbeddedNoteReferences } from "./note-embeds";
+import { useVisibleEmbedKeys } from "./use-visible-embed-keys";
 import {
   BacklinksPanel,
   EmptyEditor,
@@ -108,16 +110,11 @@ import {
 const TypeList = lazy(() => import("./TypeBrowser").then((module) => ({ default: module.TypeList })));
 const TypeInspector = lazy(() => import("./TypeBrowser").then((module) => ({ default: module.TypeInspector })));
 const TypePackBrowser = lazy(() => import("./TypeBrowser").then((module) => ({ default: module.TypePackBrowser })));
-const CodeEditor = lazy(() => import("./CodeEditor").then((module) => ({ default: module.CodeEditor })));
 const PropertiesPanel = lazy(() => import("./PropertiesPanel").then((module) => ({ default: module.PropertiesPanel })));
 const NewNoteComposer = lazy(() => import("./NewNoteComposer").then((module) => ({ default: module.NewNoteComposer })));
 const FileViewer = lazy(() => import("./FileViewer").then((module) => ({ default: module.FileViewer })));
 const FileWorkspace = lazy(() => import("./FileViewer").then((module) => ({ default: module.FileWorkspace })));
 const emptyTypeDescriptors: CollectionTypeDescriptor[] = [];
-
-function sameStringSet(current: ReadonlySet<string>, next: readonly string[]): boolean {
-  return current.size === next.length && next.every((value) => current.has(value));
-}
 
 interface Confirmation {
   title: string;
@@ -229,8 +226,8 @@ export function App({ gateway }: { gateway: CollectionGateway }) {
   const [noteSort, setNoteSort] = useState<NoteSort>(loadNoteSort);
   const [resizingPane, setResizingPane] = useState<"collection" | "list" | "inspector">();
   const [viewportWidth, setViewportWidth] = useState(() => window.innerWidth);
-  const [visibleFileEmbedKeys, setVisibleFileEmbedKeys] = useState<ReadonlySet<string>>(() => new Set());
-  const [visibleNoteEmbedKeys, setVisibleNoteEmbedKeys] = useState<ReadonlySet<string>>(() => new Set());
+  const { files: visibleFileEmbedKeys, notes: visibleNoteEmbedKeys,
+    updateFiles: updateVisibleFileEmbeds, updateNotes: updateVisibleNoteEmbeds } = useVisibleEmbedKeys(document?.path);
   const [, setSessionTick] = useState(0);
   const documentGeneration = useRef(0);
   const navigationGeneration = useRef(0);
@@ -247,16 +244,6 @@ export function App({ gateway }: { gateway: CollectionGateway }) {
   const mobileLayout = viewportWidth <= 760;
   const typeDescriptors = description?.types ?? emptyTypeDescriptors;
   const notePreviewController = useNotePreview(gateway, allNotes, typeDescriptors);
-  const updateVisibleFileEmbeds = useCallback((keys: string[]) => {
-    setVisibleFileEmbedKeys((current) => sameStringSet(current, keys) ? current : new Set(keys));
-  }, []);
-  const updateVisibleNoteEmbeds = useCallback((keys: string[]) => {
-    setVisibleNoteEmbedKeys((current) => sameStringSet(current, keys) ? current : new Set(keys));
-  }, []);
-  useEffect(() => {
-    setVisibleFileEmbedKeys(new Set());
-    setVisibleNoteEmbedKeys(new Set());
-  }, [document?.path]);
   const fileWorkspace = useFileWorkspace(
     fileAssetStore,
     fileInventory.files,
@@ -2001,44 +1988,15 @@ export function App({ gateway }: { gateway: CollectionGateway }) {
             <button className="primary-confirm-action" onClick={() => void performRename(renamePlan, true)}>Rename and update links</button>
           </div>}
           {deletePlan && deletePlan.session === noteSessions.current.active && <div className="delete-confirm" role="alert"><div><strong>Delete this note?</strong><span>{deletePlan.brokenLinkPaths.length > 0 ? `${deletePlan.brokenLinkPaths.length.toLocaleString()} ${deletePlan.brokenLinkPaths.length === 1 ? "note will keep a broken link" : "notes will keep broken links"}. ` : ""}You can undo the note deletion.</span></div><button onClick={() => setDeletePlan(undefined)}>Keep note</button><button className="danger-action" onClick={() => void deleteNote(deletePlan)}>Delete</button></div>}
-          <article className="writing-surface" style={{ "--editor-font-size": `${preferences.fontSize}px` } as CSSProperties}>
-            <label className="sr-only" htmlFor="note-title">Note title</label>
-            <input id="note-title" className="title-input" value={draft.title} onChange={(event) => changeActiveDraft((current) => ({ ...current, title: event.target.value }))} placeholder="Untitled" spellCheck="true" />
-            <Suspense fallback={<div className="body-editor code-editor-loading" role="status" aria-label="Loading note editor" aria-busy="true"><span /></div>}>
-              <CodeEditor
-                key={noteSessions.current.active?.editorSessionKey ?? document.path}
-                value={draft.body}
-                onChange={(body) => changeActiveDraft((current) => ({ ...current, body }))}
-                label="Note body"
-                language="markdown"
-                variant="writer"
-                placeholder="Start writing"
-                vimEnabled={preferences.vim}
-                lineWrapping={preferences.lineWrapping}
-                quietMarkdown={preferences.quietMarkdown}
-                autoFocus
-                className="body-editor"
-                documentId={noteSessions.current.active?.editorSessionKey}
-                currentPath={document.path}
-                recentPaths={recentPaths}
-                linkSuggestions={linkOptions}
-                linkTypes={linkTypeNames}
-                onOpenLink={navigateToNote}
-                onCreateLink={createLinkedNote}
-                onPreviewLink={notePreviewController.request}
-                onDismissLinkPreview={notePreviewController.dismiss}
-                embeddedFiles={embeddedFiles}
-                embeddedNotes={embeddedNotes}
-                onOpenFile={setOpenFileAsset}
-                files={fileInventory.files}
-                notes={allNotes}
-                onOpenFileLink={navigateToFile}
-                onVisibleFileEmbeds={updateVisibleFileEmbeds}
-                onVisibleNoteEmbeds={updateVisibleNoteEmbeds}
-                insertion={attachments.insertion}
-              />
-            </Suspense>
-          </article>
+          <MarkdownNoteEditor editorKey={noteSessions.current.active?.editorSessionKey ?? document.path}
+            draft={draft} preferences={preferences} documentId={noteSessions.current.active?.editorSessionKey}
+            currentPath={document.path} recentPaths={recentPaths} linkSuggestions={linkOptions} linkTypes={linkTypeNames}
+            embeddedFiles={embeddedFiles} embeddedNotes={embeddedNotes} files={fileInventory.files} notes={allNotes}
+            insertion={attachments.insertion} onTitleChange={(title) => changeActiveDraft((current) => ({ ...current, title }))}
+            onBodyChange={(body) => changeActiveDraft((current) => ({ ...current, body }))} onOpenLink={navigateToNote}
+            onCreateLink={createLinkedNote} onPreviewLink={notePreviewController.request}
+            onDismissLinkPreview={notePreviewController.dismiss} onOpenFile={setOpenFileAsset} onOpenFileLink={navigateToFile}
+            onVisibleFileEmbeds={updateVisibleFileEmbeds} onVisibleNoteEmbeds={updateVisibleNoteEmbeds} />
         </> : <EmptyEditor
           leadingActions={editorLeadingActions}
           notice={editorNotice}
